DPC-113: Renamed DPCWeb module to DPCAPI
Also started the renaming from application to service.
