+++ dpc-api/src/main/java/gov/cms/dpc/api/resources/AbstractJobResource.java
@@ -1,0 +1,19 @@
+package gov.cms.dpc.api.resources;
+
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.Response;
+
+@Produces("application/json")
+@Path("/Jobs")
+public abstract class AbstractJobResource {
+
+    protected AbstractJobResource() {
+        // Not used
+    }
+
+    @Path("/{jobID}")
+    @GET
+    public abstract Response checkJobStatus(String jobID);
+}
