+++ dpc-api/src/main/resources/banner.txt
@@ -1,0 +1,6 @@
+================================================================================
+
+                              DPC API Service
+
+================================================================================
+
