--- dpc-web/src/main/java/gov/cms/dpc/web/DPWebConfiguration.java
@@ -1,61 +1,0 @@
-package gov.cms.dpc.web;
-
-import ca.mestevens.java.configuration.TypesafeConfiguration;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.dropwizard.client.JerseyClientConfiguration;
-import org.hibernate.validator.constraints.NotEmpty;
-
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
-
-public class DPWebConfiguration extends TypesafeConfiguration {
-
-    private String testValue;
-    @NotEmpty
-    private String exportPath;
-
-    @Valid
-    @NotNull
-    @JsonProperty
-    private JerseyClientConfiguration httpClient = new JerseyClientConfiguration();
-
-    @NotEmpty
-    @NotNull
-    private String attributionURL;
-
-    DPWebConfiguration() {
-//        Not used;
-    }
-
-    public String getTestValue() {
-        return this.testValue;
-    }
-
-    public void setTestValue(String testValue) {
-        this.testValue = testValue;
-    }
-
-    public JerseyClientConfiguration getHttpClient() {
-        return httpClient;
-    }
-
-    public void setHttpClient(JerseyClientConfiguration httpClient) {
-        this.httpClient = httpClient;
-    }
-
-    public String getAttributionURL() {
-        return attributionURL;
-    }
-
-    public void setAttributionURL(String attributionURL) {
-        this.attributionURL = attributionURL;
-    }
-
-    public String getExportPath() {
-        return exportPath;
-    }
-
-    public void setExportPath(String exportPath) {
-        this.exportPath = exportPath;
-    }
-}
