--- dpc-web/src/main/resources/banner.txt
@@ -1,6 +1,0 @@
-================================================================================
-
-                              DPCApp
-
-================================================================================
-
